+++ connector.js
@@ -1,0 +1,7 @@
+const nullConnector = require('./src/connector/nullConnector')
+const syncConnector = require('./src/connector/syncConnector')
+
+module.exports = {
+  nullConnector,
+  syncConnector
+}